move scheduler/controller API to extras
